Refactor project structure: rename project from SPL to Axiom, update subdirectories, and adjust CMake configurations for anvil and axiom components.

+++ anvil/include/vm.h
@@ -33,15 +33,15 @@ typedef enum
 
 typedef struct
 {
-    uint16_t registers[R_COUNT];
-    uint16_t flags;
+    int registers[R_COUNT];
+    uint32_t flags;
     int ip; // Instruction Pointer
     int sp; // Stack Pointer
 } CPU;
 
 typedef struct
 {
-    uint16_t data[MEMORY_SIZE];
+    uint32_t data[MEMORY_SIZE];
 } Memory;
 
 typedef struct
